clogin passing wrong password to device when it contains a $-sign, no matter what escaping or bracing is applied in clogin, the output was enclosed in braces.
Cause appears to be due to proc find() returning a tcl list and we were
treating it as a string, which is normally fine.  However, since the value
contained a meta-character, tcl is adding the braces in what seems to be an
attempt to protect the meta-character from pre-mature expansion/eval.

Fixed by [join ""]'ing the value.  This fix was applied to all find-returned
values that have a likely-hood of containing TCL meta-characters.  ie: applied
to password, but not to autoenable or sshcmd.


git-svn-id: http://rancid.shrubbery.net/rancid/svn/rancid@1356 137ac7c6-8f6c-0410-88de-aa6437057478