Filter debugging file dlbg.txt & dlbg.txt-1 which, ATM, appears only on the ASR9k w/ IOS-XR.
Per Crisco:
This file is used for the service console feature.  It is the "active" file
from Data Logger BackGround logging.  It gets updated every minute. When it
reaches a certain file size it will be copied to the filename dlbg.txt-1 and
a new dlbg.txt will be created.   The service console is designed to log all
data every minute, regardless of whether you are looking at the interactive
service console browser or not.  The purpose of this is to have history of
"what happened" prior to some issue.  The model is that a customer could
copy this fill off the router and send to TAC for post processing at triage.
It hopefully isn't more of annoyance than a syslog for example.

Disabling the feature is problematic as it requires special privileges, but
it is not impossible.  I had to add the following to my tacacs config to be
able to disable the logging:

         service = exec {
                priv-lvl = 15
+               task = "#root-system,#cisco-support"
        }

The key is the user needs to be in the cisco-support group to be able to set
"service logger disable".


git-svn-id: http://rancid.shrubbery.net/rancid/svn/rancid@2110 137ac7c6-8f6c-0410-88de-aa6437057478